third_party/blink/web_tests/external/wpt/scroll-animations/css/scroll-timeline-in-container-query-expected.txt

This is a testharness.js-based test.
[FAIL] Timeline appearing via container queries
  assert_equals: expected "rgb(100, 100, 100)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
